backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
langchain-openai
langchain-pinecone
pinecone-client
pypdf
python-dotenv
tiktoken
